Remove the memory resource for mark duplicates
I had it at 4700 but I think 4848 is what is actually on the machine, and on a few samples I got an oom kill.  Wondering if picard uses all the memory sometime, and having mem at 4700 causes the thing to fail.  If not, I might just bump it up more.