Big update: clipOverlaps and IndelRealignment
Also, for Haplotype caller I added min base quality score and min mapping quality.  These can be set
in the config.

This gets us to angsd-ready bams, using the hard-filtered indels to guide the indel realignment.

I also edited the .test config so that it does 0
rounds of BQSR.

I made the hard-filtered indel sections output no longer temp, mostly so that I can test stuff and not have to regenerate them if I forgot to request the angsd-ready bams which will rely on those.