Remove temp() and protected() directives
CH had such a hard time getting runs to start up again because of the protected gvfs and also now that we load the genomic data bases on a chrom or chrom_group basis, it makes a lot more sense to not make the gvcf_sections as temp.  While I was at it, I figured I would not remove many of the other temp() directives.